+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 ## #.#.#
 
 ### Changed
--
+- Update serial writer to take a list of blocks as input for the chipmap instead of a file.
 
 
 ## 0.9.8