+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # CHANGELOG
 
 
-## #.#.#
+## 0.9.10
 
 ### Changed
 - Update serial writer to take a list of blocks as input for the chipmap instead of a file.